Send the from parameter onto BrowserSelection
When the usage of missingSupport was expanded to send on the path of the
source being checked and BrowserSelection was expanded to use it in #69,
it seems that MissingSupport itself never got updated to pass on the
parameter.

This was causing the process.cwd to be used as the path eventually,
which was potentially wildly inaccurate leading to the configuration not
being found.