Conform to browserslist api
Commit a1bf895ecd48b5c4ba8e82227da3c33aaafc3a48 changed the browserslist
request from
    browserslist(this.browsersRequest)
to
    browserslist(this.browsersRequest, { from })

But in browserslist the option name for the path is 'path'. This commit
corrects this. Note that this means that before, since 'path' was not specified,
browserslist would fallback to the current working directory for the
config file look up.

With this commit, if no queries are specified, browserslist will look up
for a config file up from the specified path rather than the current working
directory. This conforms with autoprefixer behavior.

Note that we only pass 'from' if it is not null, this allows
browserslist to fallback to the current directory when no path is
provided.